fixed some warnings (caused by differences of platforms)

+++ src/utils/stringutils.c
@@ -26,7 +26,7 @@ void cf_str_cleanup(cf_string_t *str) {
 int32_t cf_str_char_append(cf_string_t *str,const UChar content) {
   if(str->growth == 0) str->growth = CF_BUFSIZ;
 
-  if(str->len + sizeof(*str->content) >= str->reserved) {
+  if(str->len + (int32_t)sizeof(*str->content) >= str->reserved) {
     str->reserved += str->growth;
     str->content   = cf_alloc(str->content,(size_t)str->reserved,sizeof(*str->content),CF_ALLOC_REALLOC);
   }
@@ -43,7 +43,7 @@ int32_t cf_str_uchar32_append(cf_string_t *str,const UChar32 content) {
 
   if(str->growth == 0) str->growth = CF_BUFSIZ;
 
-  if(str->len + (U16_LENGTH(content) * sizeof(*str->content)) >= str->reserved) {
+  if(str->len + (U16_LENGTH(content) * (int32_t)sizeof(*str->content)) >= str->reserved) {
     str->reserved += str->growth;
     str->content   = cf_alloc(str->content,(size_t)str->reserved,sizeof(*str->content),CF_ALLOC_REALLOC);
   }
@@ -95,7 +95,7 @@ int32_t cf_str_cstr_set(cf_string_t *str,const UChar *content) {
 }
 
 int32_t cf_str_char_set(cf_string_t *str,const UChar *content,const int32_t length) {
-  size_t len;
+  int32_t len;
 
   if(str->growth == 0) str->growth = CF_BUFSIZ;
   len = str->growth;
